requirements.txt
greenlet==3.0.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyYAML==6.0.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
